Reportes/requirements.txt
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
